Refactored ServiceRecord to mdns.cpp it's not used elsewhere and changes to support newer mdns.h will be simpler that way.
